+++ packages/expo-age-range/CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ### 🎉 New features
 
-- [android] bump com.google.android.play:age-signals dependency to 0.0.2
+- [android] bump com.google.android.play:age-signals dependency to 0.0.2 ([#41568](https://github.com/expo/expo/pull/41568) by [@vonovak](https://github.com/vonovak))
 
 ### 🐛 Bug fixes